Fix PowerShell 7 test failures by moving SqlServerDscCI environment variable setting

+++ tests/Unit/DSC_SqlMemory.Tests.ps1
@@ -17,7 +17,7 @@ BeforeDiscovery {
             {
                 # Redirect all streams to $null, except the error stream (stream 2)
                 & "$PSScriptRoot/../../build.ps1" -Tasks 'noop' 3>&1 4>&1 5>&1 6>&1 > $null
-            }
+            }    # Set environment variable to prevent loading real SQL Server assemblies during testing}}
 
             # If the dependencies has not been resolved, this will throw an error.
             Import-Module -Name 'DscResource.Test' -Force -ErrorAction 'Stop'
@@ -32,8 +32,6 @@ BeforeDiscovery {
 BeforeAll {
     $script:dscModuleName = 'SqlServerDsc'
     $script:dscResourceName = 'DSC_SqlMemory'
-
-    $env:SqlServerDscCI = $true
 
     $script:testEnvironment = Initialize-TestEnvironment `
         -DSCModuleName $script:dscModuleName `
@@ -81,7 +79,7 @@ AfterAll {
     # Remove module common test helper.
     Get-Module -Name 'CommonTestHelper' -All | Remove-Module -Force
 
-    Remove-Item -Path 'env:SqlServerDscCI'
+    if (Test-Path -Path 'env:SqlServerDscCI') { Remove-Item -Path 'env:SqlServerDscCI' }
 }
 
 Describe 'SqlMaxDop\Get-TargetResource' -Tag 'Get' {
@@ -123,7 +121,6 @@ Describe 'SqlMaxDop\Get-TargetResource' -Tag 'Get' {
                 return New-Object -TypeName 'Microsoft.SqlServer.Management.Smo.Server' |
                     Add-Member -MemberType NoteProperty -Name 'Configuration' -Value $mockServerConfigurationObject -PassThru -Force
             }
-
 
             Mock -CommandName Connect-SQL -MockWith $mockConnectSQL
             Mock -CommandName Test-ActiveNode -MockWith {
@@ -880,7 +877,6 @@ Describe 'SqlMaxDop\Set-TargetResource' -Tag 'Set' {
                         }
                     } -PassThru -Force
             }
-
 
             Mock -CommandName Connect-SQL -MockWith $mockConnectSQL
             Mock -CommandName Test-ActiveNode -MockWith {
