+++ tests/Unit/DSC_SqlRole.Tests.ps1
@@ -17,7 +17,7 @@ BeforeDiscovery {
             {
                 # Redirect all streams to $null, except the error stream (stream 2)
                 & "$PSScriptRoot/../../build.ps1" -Tasks 'noop' 3>&1 4>&1 5>&1 6>&1 > $null
-            }
+            }    # Set environment variable to prevent loading real SQL Server assemblies during testing}}
 
             # If the dependencies has not been resolved, this will throw an error.
             Import-Module -Name 'DscResource.Test' -Force -ErrorAction 'Stop'
@@ -32,8 +32,6 @@ BeforeDiscovery {
 BeforeAll {
     $script:dscModuleName = 'SqlServerDsc'
     $script:dscResourceName = 'DSC_SqlRole'
-
-    $env:SqlServerDscCI = $true
 
     $script:testEnvironment = Initialize-TestEnvironment `
         -DSCModuleName $script:dscModuleName `
@@ -203,7 +201,7 @@ AfterAll {
     # Remove module common test helper.
     Get-Module -Name 'CommonTestHelper' -All | Remove-Module -Force
 
-    Remove-Item -Path 'env:SqlServerDscCI'
+    if (Test-Path -Path 'env:SqlServerDscCI') { Remove-Item -Path 'env:SqlServerDscCI' }
 }
 
 Describe "DSC_SqlRole\Get-TargetResource" -Tag 'Get' {
@@ -732,7 +730,6 @@ Describe "DSC_SqlRole\Set-TargetResource" -Tag 'Set' {
 
                 $mockTestParameters.Ensure         = 'Absent'
                 $mockTestParameters.ServerRoleName = 'AdminSqlForBI'
-
 
                 $mockErrorMessage = $script:localizedData.DropServerRoleSetError `
                     -f 'localhost', 'MSSQLSERVER', 'AdminSqlForBI'
