+++ tests/Unit/DSC_SqlWindowsFirewall.Tests.ps1
@@ -19,7 +19,7 @@ BeforeDiscovery {
             {
                 # Redirect all streams to $null, except the error stream (stream 2)
                 & "$PSScriptRoot/../../build.ps1" -Tasks 'noop' 3>&1 4>&1 5>&1 6>&1 > $null
-            }
+            }    # Set environment variable to prevent loading real SQL Server assemblies during testing}}
 
             # If the dependencies has not been resolved, this will throw an error.
             Import-Module -Name 'DscResource.Test' -Force -ErrorAction 'Stop'
@@ -34,8 +34,6 @@ BeforeDiscovery {
 BeforeAll {
     $script:dscModuleName = 'SqlServerDsc'
     $script:dscResourceName = 'DSC_SqlWindowsFirewall'
-
-    $env:SqlServerDscCI = $true
 
     $script:testEnvironment = Initialize-TestEnvironment `
         -DSCModuleName $script:dscModuleName `
@@ -72,7 +70,7 @@ AfterAll {
     # Remove module common test helper.
     Get-Module -Name 'CommonTestHelper' -All | Remove-Module -Force
 
-    Remove-Item -Path 'env:SqlServerDscCI'
+    if (Test-Path -Path 'env:SqlServerDscCI') { Remove-Item -Path 'env:SqlServerDscCI' }
 }
 
 Describe 'SqlWindowsFirewall\Get-TargetResource' -Tag 'Get' {
@@ -2373,7 +2371,6 @@ Describe 'SqlWindowsFirewall\Set-TargetResource' -Tag 'Set' {
             }
         }
 
-
         Context 'When using the feature ''<MockFeatures>'' and major version ''<MockSqlMajorVersion>''' -ForEach @(
             <#
                 Testing two major versions to verify Integration Services differences (i.e service name).
@@ -2565,7 +2562,7 @@ Describe 'SqlWindowsFirewall/Get-SqlRootPath' -Tag 'Helper' {
             @{
                 MockFeature = 'AS'
                 MockSqlMajorVersion = $null
-            }
+            }    # Set environment variable to prevent loading real SQL Server assemblies during testing}}
             @{
                 # Lower-case to test that casing does not matter.
                 MockFeature = 'SQLEngine'
