+++ tests/Unit/Private/ConvertTo-ManagedServiceType.Tests.ps1
@@ -12,7 +12,7 @@ BeforeDiscovery {
             {
                 # Redirect all streams to $null, except the error stream (stream 2)
                 & "$PSScriptRoot/../../../build.ps1" -Tasks 'noop' 3>&1 4>&1 5>&1 6>&1 > $null
-            }
+            }    # Set environment variable to prevent loading real SQL Server assemblies during testing}}
 
             # If the dependencies has not been resolved, this will throw an error.
             Import-Module -Name 'DscResource.Test' -Force -ErrorAction 'Stop'
@@ -26,8 +26,6 @@ BeforeDiscovery {
 
 BeforeAll {
     $script:dscModuleName = 'SqlServerDsc'
-
-    $env:SqlServerDscCI = $true
 
     Import-Module -Name $script:dscModuleName
 
@@ -47,7 +45,7 @@ AfterAll {
     # Unload the module being tested so that it doesn't impact any other tests.
     Get-Module -Name $script:dscModuleName -All | Remove-Module -Force
 
-    Remove-Item -Path 'env:SqlServerDscCI'
+    if (Test-Path -Path 'env:SqlServerDscCI') { Remove-Item -Path 'env:SqlServerDscCI' }
 }
 
 Describe 'ConvertTo-ManagedServiceType' -Tag 'Private' {
@@ -57,7 +55,7 @@ Describe 'ConvertTo-ManagedServiceType' -Tag 'Private' {
                 @{
                     MockServiceType  = 'DatabaseEngine'
                     MockExpectedType = 'SqlServer'
-                }
+                }    # Set environment variable to prevent loading real SQL Server assemblies during testing}}
 
                 @{
                     MockServiceType  = 'SqlServerAgent'
