+++ tests/Unit/Public/Set-SqlDscServerPermission.Tests.ps1
@@ -11,7 +11,7 @@ BeforeDiscovery {
             {
                 # Redirect all streams to $null, except the error stream (stream 2)
                 & "$PSScriptRoot/../../../build.ps1" -Tasks 'noop' 3>&1 4>&1 5>&1 6>&1 > $null
-            }
+            }    # Set environment variable to prevent loading real SQL Server assemblies during testing}}
 
             # If the dependencies has not been resolved, this will throw an error.
             Import-Module -Name 'DscResource.Test' -Force -ErrorAction 'Stop'
@@ -25,8 +25,6 @@ BeforeDiscovery {
 
 BeforeAll {
     $script:dscModuleName = 'SqlServerDsc'
-
-    $env:SqlServerDscCI = $true
 
     Import-Module -Name $script:dscModuleName
 
@@ -46,7 +44,7 @@ AfterAll {
     # Unload the module being tested so that it doesn't impact any other tests.
     Get-Module -Name $script:dscModuleName -All | Remove-Module -Force
 
-    Remove-Item -Path 'env:SqlServerDscCI'
+    if (Test-Path -Path 'env:SqlServerDscCI') { Remove-Item -Path 'env:SqlServerDscCI' }
 }
 
 Describe 'Set-SqlDscServerPermission' -Tag 'Public' {
@@ -125,7 +123,6 @@ Describe 'Set-SqlDscServerPermission' -Tag 'Public' {
                         $script:mockMethodDenyCallCount += 1
                     } -PassThru -Force
 
-
                 Mock -CommandName Test-SqlDscIsLogin -MockWith {
                     return $true
                 }
@@ -162,7 +159,6 @@ Describe 'Set-SqlDscServerPermission' -Tag 'Public' {
                     Add-Member -MemberType 'ScriptMethod' -Name 'Deny' -Value {
                         $script:mockMethodDenyCallCount += 1
                     } -PassThru -Force
-
 
                 Mock -CommandName Test-SqlDscIsLogin -MockWith {
                     return $true
@@ -261,7 +257,6 @@ Describe 'Set-SqlDscServerPermission' -Tag 'Public' {
                             $script:mockMethodGrantUsingWithGrantCallCount += 1
                         }
                     } -PassThru -Force
-
 
                 Mock -CommandName Test-SqlDscIsLogin -MockWith {
                     return $true
@@ -527,7 +522,6 @@ Describe 'Set-SqlDscServerPermission' -Tag 'Public' {
                         $script:mockMethodDenyCallCount += 1
                     } -PassThru -Force
 
-
                 Mock -CommandName Test-SqlDscIsLogin -MockWith {
                     return $false
                 }
@@ -564,7 +558,6 @@ Describe 'Set-SqlDscServerPermission' -Tag 'Public' {
                     Add-Member -MemberType 'ScriptMethod' -Name 'Deny' -Value {
                         $script:mockMethodDenyCallCount += 1
                     } -PassThru -Force
-
 
                 Mock -CommandName Test-SqlDscIsLogin -MockWith {
                     return $false
@@ -663,7 +656,6 @@ Describe 'Set-SqlDscServerPermission' -Tag 'Public' {
                             $script:mockMethodGrantUsingWithGrantCallCount += 1
                         }
                     } -PassThru -Force
-
 
                 Mock -CommandName Test-SqlDscIsLogin -MockWith {
                     return $false
